+++ src/main/java/com/hubspot/jinjava/lib/tag/eager/EagerIncludeTag.java
@@ -3,8 +3,6 @@ package com.hubspot.jinjava.lib.tag.eager;
 import com.google.common.annotations.Beta;
 import com.hubspot.jinjava.interpret.JinjavaInterpreter;
 import com.hubspot.jinjava.lib.tag.IncludeTag;
-import com.hubspot.jinjava.lib.tag.eager.importing.EagerImportingStrategyFactory;
-import com.hubspot.jinjava.loader.RelativePathResolver;
 import com.hubspot.jinjava.tree.TagNode;
 import com.hubspot.jinjava.util.EagerReconstructionUtils;
 import com.hubspot.jinjava.util.HelperStringTokenizer;
@@ -30,14 +28,11 @@ public class EagerIncludeTag extends EagerTagDecorator<IncludeTag> {
         tagNode.getStartPosition()
       );
       templateFile = interpreter.resolveResourceLocation(templateFile);
-      final String initialPathSetter =
-        EagerImportingStrategyFactory.getSetTagForCurrentPath(interpreter);
-      final String newPathSetter = EagerReconstructionUtils.buildBlockOrInlineSetTag(
-        RelativePathResolver.CURRENT_PATH_CONTEXT_KEY,
+      return EagerReconstructionUtils.wrapPathAroundText(
+        output,
         templateFile,
         interpreter
       );
-      return newPathSetter + output + initialPathSetter;
     }
     return output;
   }