+++ src/main/java/com/hubspot/jinjava/lib/tag/eager/EagerSetTagStrategy.java
@@ -3,8 +3,9 @@ package com.hubspot.jinjava.lib.tag.eager;
 import com.google.common.annotations.Beta;
 import com.hubspot.jinjava.interpret.DeferredValueException;
 import com.hubspot.jinjava.interpret.JinjavaInterpreter;
+import com.hubspot.jinjava.interpret.MetaContextVariables;
 import com.hubspot.jinjava.lib.tag.SetTag;
-import com.hubspot.jinjava.lib.tag.eager.importing.AliasedEagerImportingStrategy;
+import com.hubspot.jinjava.loader.RelativePathResolver;
 import com.hubspot.jinjava.tree.TagNode;
 import com.hubspot.jinjava.util.EagerReconstructionUtils;
 import com.hubspot.jinjava.util.PrefixToPreserveState;
@@ -13,6 +14,7 @@ import java.util.List;
 import java.util.Optional;
 import java.util.StringJoiner;
 import java.util.stream.Collectors;
+import java.util.stream.Stream;
 import org.apache.commons.lang3.tuple.Triple;
 
 @Beta
@@ -52,9 +54,14 @@ public abstract class EagerSetTagStrategy {
       expression,
       interpreter
     );
+    boolean triedResolve = false;
     if (
       eagerExecutionResult.getResult().isFullyResolved() &&
-      !interpreter.getContext().isDeferredExecutionMode()
+      !interpreter.getContext().isDeferredExecutionMode() &&
+      (Arrays
+          .stream(variables)
+          .noneMatch(RelativePathResolver.CURRENT_PATH_CONTEXT_KEY::equals) ||
+        interpreter.getContext().getPenultimateParent().getDeferredTokens().isEmpty()) // Prevents set tags from disappearing in nested interpretation
     ) {
       EagerReconstructionUtils.commitSpeculativeBindings(
         interpreter,
@@ -66,6 +73,7 @@ public abstract class EagerSetTagStrategy {
         eagerExecutionResult,
         interpreter
       );
+      triedResolve = true;
       if (maybeResolved.isPresent()) {
         return maybeResolved.get();
       }
@@ -76,10 +84,7 @@ public abstract class EagerSetTagStrategy {
       eagerExecutionResult,
       interpreter
     );
-    if (
-      eagerExecutionResult.getResult().isFullyResolved() &&
-      interpreter.getContext().isDeferredExecutionMode()
-    ) {
+    if (eagerExecutionResult.getResult().isFullyResolved() && !triedResolve) {
       attemptResolve(tagNode, variables, eagerExecutionResult, interpreter);
     }
     return buildImage(tagNode, variables, eagerExecutionResult, triple, interpreter);
@@ -155,39 +160,66 @@ public abstract class EagerSetTagStrategy {
   }
 
   public static String getSuffixToPreserveState(
-    String variables,
+    List<String> varList,
     JinjavaInterpreter interpreter
   ) {
-    if (variables.isEmpty()) {
+    if (varList.isEmpty()) {
       return "";
     }
-    StringBuilder suffixToPreserveState = new StringBuilder();
-    Optional<String> maybeTemporaryImportAlias =
-      AliasedEagerImportingStrategy.getTemporaryImportAlias(interpreter.getContext());
-    if (
-      maybeTemporaryImportAlias.isPresent() &&
-      !AliasedEagerImportingStrategy.isTemporaryImportAlias(variables) &&
-      !interpreter.getContext().getComputedMetaContextVariables().contains(variables)
-    ) {
-      if (!interpreter.getContext().containsKey(maybeTemporaryImportAlias.get())) {
-        if (
-          interpreter.retraceVariable(
-            String.format(
-              "%s.%s",
-              interpreter.getContext().getImportResourceAlias().get(),
-              variables
-            ),
-            -1
-          ) !=
-          null
-        ) {
-          throw new DeferredValueException(
-            "Cannot modify temporary import alias outside of import tag"
-          );
-        }
-      }
-      String updateString = getUpdateString(variables);
+    return getSuffixToPreserveState(varList.stream(), interpreter);
+  }
 
+  public static String getSuffixToPreserveState(
+    String[] varList,
+    JinjavaInterpreter interpreter
+  ) {
+    if (varList.length == 0) {
+      return "";
+    }
+    return getSuffixToPreserveState(Arrays.stream(varList), interpreter);
+  }
+
+  private static String getSuffixToPreserveState(
+    Stream<String> varStream,
+    JinjavaInterpreter interpreter
+  ) {
+    StringBuilder suffixToPreserveState = new StringBuilder();
+    Optional<String> maybeTemporaryImportAlias = interpreter
+      .getContext()
+      .getImportResourceAlias()
+      .map(MetaContextVariables::getTemporaryImportAlias);
+    if (maybeTemporaryImportAlias.isPresent()) {
+      boolean stillInsideImportTag = interpreter
+        .getContext()
+        .containsKey(maybeTemporaryImportAlias.get());
+      List<String> filteredVars = varStream
+        .filter(var ->
+          !MetaContextVariables.isMetaContextVariable(var, interpreter.getContext())
+        )
+        .peek(var -> {
+          if (!stillInsideImportTag) {
+            if (
+              interpreter.retraceVariable(
+                String.format(
+                  "%s.%s",
+                  interpreter.getContext().getImportResourceAlias().get(),
+                  var
+                ),
+                -1
+              ) !=
+              null
+            ) {
+              throw new DeferredValueException(
+                "Cannot modify temporary import alias outside of import tag"
+              );
+            }
+          }
+        })
+        .collect(Collectors.toList());
+      if (filteredVars.isEmpty()) {
+        return "";
+      }
+      String updateString = getUpdateString(filteredVars);
       // Don't need to render because the temporary import alias's value is always deferred, and rendering will do nothing
       suffixToPreserveState.append(
         EagerReconstructionUtils.buildDoUpdateTag(
@@ -197,15 +229,10 @@ public abstract class EagerSetTagStrategy {
         )
       );
     }
-
     return suffixToPreserveState.toString();
   }
 
-  private static String getUpdateString(String variables) {
-    List<String> varList = Arrays
-      .stream(variables.split(","))
-      .map(String::trim)
-      .collect(Collectors.toList());
+  private static String getUpdateString(List<String> varList) {
     StringJoiner updateString = new StringJoiner(",");
     // Update the alias map to the value of the set variable.
     varList.forEach(var -> updateString.add(String.format("'%s': %s", var, var)));