+++ src/main/java/com/hubspot/jinjava/lib/tag/eager/EagerTagFactory.java
@@ -5,7 +5,9 @@ import com.google.common.base.Throwables;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
 import com.hubspot.jinjava.lib.tag.BlockTag;
+import com.hubspot.jinjava.lib.tag.BreakTag;
 import com.hubspot.jinjava.lib.tag.CallTag;
+import com.hubspot.jinjava.lib.tag.ContinueTag;
 import com.hubspot.jinjava.lib.tag.CycleTag;
 import com.hubspot.jinjava.lib.tag.DoTag;
 import com.hubspot.jinjava.lib.tag.ElseIfTag;
@@ -54,6 +56,8 @@ public class EagerTagFactory {
     .add(ElseTag.class)
     .add(RawTag.class)
     .add(ExtendsTag.class) // TODO support reconstructing extends tags
+    .add(BreakTag.class) // TODO support eager break tag
+    .add(ContinueTag.class) // TODO support eager continue tag
     .build();
 
   @SuppressWarnings("unchecked")