+++ src/main/java/com/hubspot/jinjava/loader/RelativePathResolver.java
@@ -11,13 +11,7 @@ public class RelativePathResolver implements LocationResolver {
   @Override
   public String resolve(String path, JinjavaInterpreter interpreter) {
     if (path.startsWith("./") || path.startsWith("../")) {
-      String parentPath = interpreter
-        .getContext()
-        .getCurrentPathStack()
-        .peek()
-        .orElseGet(() ->
-          (String) interpreter.getContext().getOrDefault(CURRENT_PATH_CONTEXT_KEY, "")
-        );
+      String parentPath = getCurrentPathFromStackOrKey(interpreter);
 
       Path templatePath = Paths.get(parentPath);
       Path folderPath = templatePath.getParent() != null
@@ -28,5 +22,15 @@ public class RelativePathResolver implements LocationResolver {
       }
     }
     return path;
+  }
+
+  public static String getCurrentPathFromStackOrKey(JinjavaInterpreter interpreter) {
+    return interpreter
+      .getContext()
+      .getCurrentPathStack()
+      .peek()
+      .orElseGet(() ->
+        (String) interpreter.getContext().getOrDefault(CURRENT_PATH_CONTEXT_KEY, "")
+      );
   }
 }